Workaround for spawnProcess not working on OSX
On OSX, there is a bug that if the stdin is busy being read, then the
std.process.execute function will be blocked until the read is
fulfilled. This happen because both the reading and the subprocess
spawning will issue a flock on the stdin file.
This workaround will replace the global stdin with a File pointing to
/dev/null while maintaning a reference to the original stdin in the
app.d scope.