stdin and stdout are only available on app.d
The stdin and stdout are replaced respectively by /dev/null and stderr. The original Files are only accessible in the app.d file. This ensures that no other module or subprocess will be able to write in these File, which is important as they are being used to do the communication with the parent process.