fix unescapeString call
the argument should already be a string literal, but I couldn't
reproduce the issue yet. So now in release mode there is a warning when
the contract would fail and in debug mode a regular exception is thrown.

The malformed string literal should be fixed with unescape string, but
it might just be memory corruption as well.